Refactor unset_endpoint_sender_pid, remove signal sender pid logic.
We dispatch "gp_operate_endpoints_token" UDF by
"CdbDispatchCommandToSegments" and "CdbDispatchCommand",
It'll always dispatch to writer gang, which is the gang that
allocate endpoints on. Since we always allocate endpoints on
the top most slice gang. This will make sure "unset_endpoint_sender_pid"
is called on the right QE in every situation. So we don't need to
consider signal to right QE anymore.