Latch to wait on write gang for query starting
The endpoints allocation is relied on the UDF command dispatch, then the
query started by PortalRun. This is roughly what DECLARE PARALLEL CURSOR
does.

One thing we need to ensure is that the endpoint has to be ready for
retrieve sessions. But returning of PortalRun cannot ensure that.
PortalRun just starts the query on read gang which is a different process
from the command dispatch target (write gang).

This commit tries to address this issue by introduing as shared latch on
QE:

QD: Create Parallel Cursor
  QE(write gang): Alloc endpoints and create the init_done latch
  QE(read gang): Receieve PortalRun request, start query, create dest
                 receiver, set init_done latch.
  QE(write gang): Wait for the init_done latch.

Parallel Cursor declaration finish.

Authored-by: Chen Mulong <muchen@pivotal.io>