Do dsm_detach when receiving finished
The dsm_detach will send SIGUSR1 which will interrupt the procLatch.
That causes problems if the sender is waiting to finish.
So do dsm_detach as soon as the receiving finishes. Thus sender will
start wait_parallel_retrieve_close() only when the receiver's dsm
detached.

And the retrieve_quit test has been fixed.

Authored-by: Chen Mulong <muchen@pivotal.io>